Expand UI tests for GPS
